changes for rf acceleration
preserve precision of printouts

changes needed to make acceleration work within a Synergia simulation.
The beamline RF cavitites must not be turned off for registration.
Lattice::get_reference_particle() has to return a pointer so the energy
can be modified.
Added tests for acceleration.

change saved scalar pz in diagnostics to a 1d array so it follows acceleration.  analysis tools syn*plot adjusted so that it knows how to deal with it.