Adjust CMakeLists.txt for Boost 1.71
Boost 1.71 has its own CMake support that differs from the
support in recent CMake versions which we use. Since the
CMake-supplied support for Boost works with the many different
version of Boost we use, it seems easier to rely on CMake's
support of Boost, rather than Boost's support of CMake.