Adjust both macOS CI builds to fix CI failures
A recent update to macOS makes it such that we now need to pip-install mpi4py, rather using the one from Homebrew.
In addition to mpi4py, we now also pip-install (or update) setuptools and wheel, to avoid possible troubles with the building of mpi4py.
This commit also adjusts the use of NInja to allow it to choose how many processes to run.